fix(ci): MSVC C4365 — xpe_verify_metrics compute_flatness hist 인덱스 수정
hist[bin]++에서 int bin을 std::vector::operator[]에 전달하여
signed/unsigned 불일치 경고 발생. static_cast<size_t>(bin) 적용.

+++ modules/preprocess/src/xpe_verify_metrics.cpp
@@ -84,7 +84,7 @@ namespace {
         for (double v : values) {
             int bin = static_cast<int>((v - min_val) / bin_width);
             if (bin >= bins) bin = bins - 1;
-            hist[bin]++;
+            hist[static_cast<size_t>(bin)]++;
         }
 
         // Compute entropy
